Big cleanup.

+++ dependencies/SRC/database/Field.h
@@ -14,11 +14,11 @@ public:
     RONIN_INLINE float GetFloat() { return mValue ? static_cast<float>(atof(mValue)) : 0; }
     RONIN_INLINE bool GetBool() { return mValue ? atoi(mValue) > 0 : false; }
     RONIN_INLINE uint8 GetUInt8() { return mValue ? static_cast<uint8>(atol(mValue)) : 0; }
-    RONIN_INLINE int8 GetInt8() { return mValue ? static_cast<int8>(atol(mValue)) : 0; }
+    RONIN_INLINE int8 GetInt8() { return mValue ? static_cast<int8>(atoi(mValue)) : 0; }
     RONIN_INLINE uint16 GetUInt16() { return mValue ? static_cast<uint16>(atol(mValue)) : 0; }
-    RONIN_INLINE int16 GetInt16() { return mValue ? static_cast<int16>(atol(mValue)) : 0; }
+    RONIN_INLINE int16 GetInt16() { return mValue ? static_cast<int16>(atoi(mValue)) : 0; }
     RONIN_INLINE uint32 GetUInt32() { return mValue ? static_cast<uint32>(atol(mValue)) : 0; }
-    RONIN_INLINE int32 GetInt32() { return mValue ? static_cast<int32>(atol(mValue)) : 0; }
+    RONIN_INLINE int32 GetInt32() { return mValue ? static_cast<int32>(atoi(mValue)) : 0; }
     uint64 GetUInt64() 
     {
         if(mValue)
